refactor: move transaction logic to ApiKineticService
