+++ libs/api/account/data-access/src/lib/api-account-data-access.module.ts
@@ -1,12 +1,17 @@
 import { ApiAppDataAccessModule } from '@kin-kinetic/api/app/data-access'
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
-import { ApiSolanaDataAccessModule } from '@kin-kinetic/api/solana/data-access'
+import { ApiKineticDataAccessModule } from '@kin-kinetic/api/kinetic/data-access'
 import { ApiTransactionDataAccessModule } from '@kin-kinetic/api/transaction/data-access'
 import { Module } from '@nestjs/common'
 import { ApiAccountDataAccessService } from './api-account-data-access.service'
 
 @Module({
-  imports: [ApiAppDataAccessModule, ApiCoreDataAccessModule, ApiSolanaDataAccessModule, ApiTransactionDataAccessModule],
+  imports: [
+    ApiAppDataAccessModule,
+    ApiCoreDataAccessModule,
+    ApiKineticDataAccessModule,
+    ApiTransactionDataAccessModule,
+  ],
   providers: [ApiAccountDataAccessService],
   exports: [ApiAccountDataAccessService],
 })
