+++ libs/api/account/data-access/src/lib/api-account-data-access.service.spec.ts
@@ -1,6 +1,6 @@
 import { ApiAppDataAccessModule } from '@kin-kinetic/api/app/data-access'
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
-import { ApiSolanaDataAccessModule } from '@kin-kinetic/api/solana/data-access'
+import { ApiKineticDataAccessModule } from '@kin-kinetic/api/kinetic/data-access'
 import { ApiTransactionDataAccessModule } from '@kin-kinetic/api/transaction/data-access'
 import { Test } from '@nestjs/testing'
 import { ApiAccountDataAccessService } from './api-account-data-access.service'
@@ -13,7 +13,7 @@ describe('ApiAccountDataAccessService', () => {
       imports: [
         ApiAppDataAccessModule,
         ApiCoreDataAccessModule,
-        ApiSolanaDataAccessModule,
+        ApiKineticDataAccessModule,
         ApiTransactionDataAccessModule,
       ],
       providers: [ApiAccountDataAccessService],
