+++ libs/api/airdrop/data-access/src/lib/api-airdrop-data-access.module.ts
@@ -1,11 +1,11 @@
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
-import { ApiSolanaDataAccessModule } from '@kin-kinetic/api/solana/data-access'
+import { ApiKineticDataAccessModule } from '@kin-kinetic/api/kinetic/data-access'
 import { Module } from '@nestjs/common'
 import { ApiAirdropDataAccessService } from './api-airdrop-data-access.service'
 
 @Module({
   providers: [ApiAirdropDataAccessService],
   exports: [ApiAirdropDataAccessService],
-  imports: [ApiCoreDataAccessModule, ApiSolanaDataAccessModule],
+  imports: [ApiCoreDataAccessModule, ApiKineticDataAccessModule],
 })
 export class ApiAirdropDataAccessModule {}
