+++ libs/api/airdrop/data-access/src/lib/api-airdrop-data-access.service.spec.ts
@@ -1,5 +1,5 @@
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
-import { ApiSolanaDataAccessModule } from '@kin-kinetic/api/solana/data-access'
+import { ApiKineticDataAccessModule } from '@kin-kinetic/api/kinetic/data-access'
 import { Test } from '@nestjs/testing'
 import { ApiAirdropDataAccessService } from './api-airdrop-data-access.service'
 
@@ -8,7 +8,7 @@ describe('ApiAirdropDataAccessService', () => {
 
   beforeEach(async () => {
     const module = await Test.createTestingModule({
-      imports: [ApiCoreDataAccessModule, ApiSolanaDataAccessModule],
+      imports: [ApiCoreDataAccessModule, ApiKineticDataAccessModule],
       providers: [ApiAirdropDataAccessService],
     }).compile()
