+++ libs/api/kinetic/data-access/src/index.ts
@@ -1,2 +1,4 @@
-export * from './lib/api-kinetic.service'
 export * from './lib/api-kinetic-data-access.module'
+export * from './lib/api-kinetic.service'
+export * from './lib/interfaces/process-transaction-options'
+export * from './lib/interfaces/transaction-with-errors'
