+++ libs/api/kinetic/data-access/src/lib/api-kinetic-data-access.module.ts
@@ -1,10 +1,12 @@
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
+import { ApiSolanaDataAccessModule } from '@kin-kinetic/api/solana/data-access'
+import { ApiWebhookDataAccessModule } from '@kin-kinetic/api/webhook/data-access'
 import { Module } from '@nestjs/common'
 import { ApiKineticService } from './api-kinetic.service'
 
 @Module({
   exports: [ApiKineticService],
-  imports: [ApiCoreDataAccessModule],
+  imports: [ApiCoreDataAccessModule, ApiSolanaDataAccessModule, ApiWebhookDataAccessModule],
   providers: [ApiKineticService],
 })
 export class ApiKineticDataAccessModule {}
