+++ libs/api/kinetic/data-access/src/lib/api-kinetic.service.spec.ts
@@ -1,4 +1,6 @@
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
+import { ApiSolanaDataAccessModule } from '@kin-kinetic/api/solana/data-access'
+import { ApiWebhookDataAccessModule } from '@kin-kinetic/api/webhook/data-access'
 import { Test } from '@nestjs/testing'
 import { ApiKineticService } from './api-kinetic.service'
 
@@ -7,7 +9,7 @@ describe('ApiKineticDataAccessService', () => {
 
   beforeEach(async () => {
     const module = await Test.createTestingModule({
-      imports: [ApiCoreDataAccessModule],
+      imports: [ApiCoreDataAccessModule, ApiSolanaDataAccessModule, ApiWebhookDataAccessModule],
       providers: [ApiKineticService],
     }).compile()
