+++ libs/api/kinetic/data-access/src/lib/interfaces/transaction-with-errors.ts
@@ -1,0 +1,3 @@
+import { Transaction, TransactionError } from '@prisma/client'
+
+export type TransactionWithErrors = Transaction & { errors: TransactionError[] }
