+++ libs/api/kinetic/util/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/api-kinetic.helpers'
+export * from './lib/parse-transaction-error'
