--- libs/api/kinetic/util/src/lib/api-kinetic.helpers.ts
@@ -1,3 +1,0 @@
-export function foo(): string {
-  return 'bar'
-}
