+++ libs/api/transaction/data-access/src/lib/api-transaction-data-access.module.ts
@@ -1,12 +1,12 @@
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
-import { ApiSolanaDataAccessModule } from '@kin-kinetic/api/solana/data-access'
-import { ApiWebhookDataAccessModule } from '@kin-kinetic/api/webhook/data-access'
+import { ApiKineticDataAccessModule } from '@kin-kinetic/api/kinetic/data-access'
 import { Module } from '@nestjs/common'
+
 import { ApiTransactionDataAccessService } from './api-transaction-data-access.service'
 import { ApiTransactionUserDataAccessService } from './api-transaction-user-data-access.service'
 
 @Module({
-  imports: [ApiCoreDataAccessModule, ApiSolanaDataAccessModule, ApiWebhookDataAccessModule],
+  imports: [ApiCoreDataAccessModule, ApiKineticDataAccessModule],
   providers: [ApiTransactionDataAccessService, ApiTransactionUserDataAccessService],
   exports: [ApiTransactionDataAccessService, ApiTransactionUserDataAccessService],
 })
