+++ libs/api/transaction/data-access/src/lib/api-transaction-data-access.service.spec.ts
@@ -1,6 +1,5 @@
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
-import { ApiSolanaDataAccessModule } from '@kin-kinetic/api/solana/data-access'
-import { ApiWebhookDataAccessModule } from '@kin-kinetic/api/webhook/data-access'
+import { ApiKineticDataAccessModule } from '@kin-kinetic/api/kinetic/data-access'
 import { Test } from '@nestjs/testing'
 import { ApiTransactionDataAccessService } from './api-transaction-data-access.service'
 
@@ -9,7 +8,7 @@ describe('ApiTransactionDataAccessService', () => {
 
   beforeEach(async () => {
     const module = await Test.createTestingModule({
-      imports: [ApiCoreDataAccessModule, ApiSolanaDataAccessModule, ApiWebhookDataAccessModule],
+      imports: [ApiCoreDataAccessModule, ApiKineticDataAccessModule],
       providers: [ApiTransactionDataAccessService],
     }).compile()
