+++ libs/api/wallet/data-access/src/lib/api-wallet-data-access.module.ts
@@ -1,5 +1,5 @@
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
-import { ApiSolanaDataAccessModule } from '@kin-kinetic/api/solana/data-access'
+import { ApiKineticDataAccessModule } from '@kin-kinetic/api/kinetic/data-access'
 import { ApiWebhookDataAccessModule } from '@kin-kinetic/api/webhook/data-access'
 import { Module } from '@nestjs/common'
 import { ApiWalletAdminDataAccessService } from './api-wallet-admin-data-access.service'
@@ -8,6 +8,6 @@ import { ApiWalletUserDataAccessService } from './api-wallet-user-data-access.se
 @Module({
   providers: [ApiWalletAdminDataAccessService, ApiWalletUserDataAccessService],
   exports: [ApiWalletAdminDataAccessService, ApiWalletUserDataAccessService],
-  imports: [ApiCoreDataAccessModule, ApiSolanaDataAccessModule, ApiWebhookDataAccessModule],
+  imports: [ApiCoreDataAccessModule, ApiKineticDataAccessModule, ApiWebhookDataAccessModule],
 })
 export class ApiWalletDataAccessModule {}
