+++ libs/api/wallet/data-access/src/lib/api-wallet-data-access.service.spec.ts
@@ -1,5 +1,5 @@
 import { ApiCoreDataAccessModule } from '@kin-kinetic/api/core/data-access'
-import { ApiSolanaDataAccessModule } from '@kin-kinetic/api/solana/data-access'
+import { ApiKineticDataAccessModule } from '@kin-kinetic/api/kinetic/data-access'
 import { ApiWebhookDataAccessModule } from '@kin-kinetic/api/webhook/data-access'
 import { Test } from '@nestjs/testing'
 import { ApiWalletUserDataAccessService } from './api-wallet-user-data-access.service'
@@ -9,7 +9,7 @@ describe('ApiWalletDataAccessService', () => {
 
   beforeEach(async () => {
     const module = await Test.createTestingModule({
-      imports: [ApiCoreDataAccessModule, ApiSolanaDataAccessModule, ApiWebhookDataAccessModule],
+      imports: [ApiCoreDataAccessModule, ApiKineticDataAccessModule, ApiWebhookDataAccessModule],
       providers: [ApiWalletUserDataAccessService],
     }).compile()
